requirements.txt
loguru==0.7.2
einops==0.6.0
mmcv==1.2.5
scipy==1.9.1
pytorch-warmup==0.1.1
imgaug==0.4.0
albumentations==0.4.6
h5py==3.8.0
scikit-learn==1.3.2
p_tqdm==1.4.0
pathspec==0.12.1
spikingjelly==0.0.0.12
